bug fixing

+++ src/main/java/com/example/user_management_service/model/Medicine.java
@@ -57,7 +57,7 @@ public class Medicine {
 
     private Integer quantity;
 
-    private Integer prescription;
+    private Double prescription;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "volume")
